added consistency check for authors with highly similar names
